apiserver/requirements.txt
attrs>=22.1.0,<23
azure-storage-blob>=12.13.1
bcrypt>=3.1.4
boltons>=19.1.0
boto3>=1.26
boto3-stubs[s3]>=1.26
clearml>=1.10.3
dpath>=1.4.2,<2.0
elasticsearch==8.17.0
fastjsonschema>=2.8
flask-compress>=1.4.0
flask-cors>=3.0.5
flask>=2.3.3
furl>=2.0.0
google-cloud-storage>=2.8.0
gunicorn>=23.0.0
humanfriendly>=4.17
jinja2
jsonmodels>=2.3
jsonschema>=2.6.0
luqum>=0.10.0
mongoengine==0.29.1
nested_dict>=1.61
packaging==20.3
pillow>=10.3.0  # fix vulnerability derived from clearml 1.18.0
psutil>=5.6.5
pyhocon>=0.3.35r
pyjwt>=2.4.0
pymongo==4.10.1
python-rapidjson>=0.6.3
redis==5.2.1
# requests>=2.13.0
# requests<=2.31.0
requests
semantic_version>=2.8.3,<3
# setuptools>=78.1.1
setuptools==68.2.0
six
validators>=0.12.4
urllib3>=1.26.18
werkzeug>=3.0.1

#Added
casdoor>=1.36.0
aiohttp